Refactor WWW class to add temp_file property and simplify download_binary method

+++ src/utils_future/WWW.py
@@ -1,6 +1,11 @@
+import hashlib
+import os
+import tempfile
+
 import requests
 
 from utils_future.console.Log import Log
+from utils_future.file.File import File
 
 log = Log("WWW")
 
@@ -23,6 +28,14 @@ class WWW:
     def __str__(self) -> str:
         return f"🌐{self.url}"
 
+    @property
+    def temp_file(self):
+        temp_dir = os.path.join(tempfile.gettempdir(), "lk_dcs", "www")
+        os.makedirs(temp_dir, exist_ok=True)
+        h = hashlib.md5(self.url.encode("utf-8")).hexdigest()
+        ext = self.url.split(".")[-1]
+        return File(os.path.join(temp_dir, f"{h}.{ext}"))
+
     def get_response(self):
         response = requests.get(
             self.url,
@@ -33,10 +46,14 @@ class WWW:
         response.raise_for_status()
         return response
 
-    def download_binary(self, file_path) -> str:
+    def download_binary(self) -> str:
+        if self.temp_file.exists:
+            return self.temp_file
+
         CHUNK_SIZE = 1024
         response = self.get_response()
+        file_path = self.temp_file.path
         with open(file_path, "wb") as fd:
             for chunk in response.iter_content(CHUNK_SIZE):
                 fd.write(chunk)
-        return file_path
+        return self.temp_file
